move handler code from common into a separate module
